Switched shader-ids to s8, and render batch size to u8, which saves some memory. I highly doubt we'd ever need 100 shaders! CLOC: 12801

+++ src/assets.h
@@ -32,7 +32,7 @@ struct Cursor
 
 struct Shader
 {
-	s32 rendererShaderID;
+	s8 rendererShaderID;
 
 	String vertexShader;
 	String fragmentShader;
